feat: move Tor settings model and ViewModel to commons (KMP)
Move TorSettings, TorType, TorPresetType, TorSettingsFlow, and
TorDialogViewModel to commons/tor package for cross-platform use.

- Strip Android R.string references from enum constructors
- Add extension properties for Android resource IDs in app module
- Backward-compatible typealiases in original package
- Replace Java Integer.parseInt with Kotlin toInt()
- Build verified: commons:compileKotlinJvm + amethyst:compilePlayDebugKotlin
