+++ amethyst/src/main/java/com/vitorpamplona/amethyst/ui/tor/TorSettingsFlow.kt
@@ -20,155 +20,5 @@
  */
 package com.vitorpamplona.amethyst.ui.tor
 
-import androidx.compose.runtime.Stable
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.combine
-
-@Stable
-class TorSettingsFlow(
-    val torType: MutableStateFlow<TorType> = MutableStateFlow(TorType.INTERNAL),
-    val externalSocksPort: MutableStateFlow<Int> = MutableStateFlow(9050),
-    val onionRelaysViaTor: MutableStateFlow<Boolean> = MutableStateFlow(true),
-    val dmRelaysViaTor: MutableStateFlow<Boolean> = MutableStateFlow(true),
-    val newRelaysViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val trustedRelaysViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val urlPreviewsViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val profilePicsViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val imagesViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val videosViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val moneyOperationsViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val nip05VerificationsViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-    val mediaUploadsViaTor: MutableStateFlow<Boolean> = MutableStateFlow(false),
-) {
-    // emits at every change in any of the properties.
-    val propertyWatchFlow =
-        combine<Any, TorSettings>(
-            listOf(
-                torType,
-                externalSocksPort,
-                onionRelaysViaTor,
-                dmRelaysViaTor,
-                newRelaysViaTor,
-                trustedRelaysViaTor,
-                urlPreviewsViaTor,
-                profilePicsViaTor,
-                imagesViaTor,
-                videosViaTor,
-                moneyOperationsViaTor,
-                nip05VerificationsViaTor,
-                mediaUploadsViaTor,
-            ),
-        ) { flows ->
-            TorSettings(
-                flows[0] as TorType,
-                flows[1] as Int,
-                flows[2] as Boolean,
-                flows[3] as Boolean,
-                flows[4] as Boolean,
-                flows[5] as Boolean,
-                flows[6] as Boolean,
-                flows[7] as Boolean,
-                flows[8] as Boolean,
-                flows[9] as Boolean,
-                flows[10] as Boolean,
-                flows[11] as Boolean,
-                flows[12] as Boolean,
-            )
-        }
-
-    fun toSettings(): TorSettings =
-        TorSettings(
-            torType.value,
-            externalSocksPort.value,
-            onionRelaysViaTor.value,
-            dmRelaysViaTor.value,
-            newRelaysViaTor.value,
-            trustedRelaysViaTor.value,
-            urlPreviewsViaTor.value,
-            profilePicsViaTor.value,
-            imagesViaTor.value,
-            videosViaTor.value,
-            moneyOperationsViaTor.value,
-            nip05VerificationsViaTor.value,
-            mediaUploadsViaTor.value,
-        )
-
-    fun update(torSettings: TorSettings): Boolean {
-        var any = false
-
-        if (torType.value != torSettings.torType) {
-            torType.tryEmit(torSettings.torType)
-            any = true
-        }
-        if (externalSocksPort.value != torSettings.externalSocksPort) {
-            externalSocksPort.tryEmit(torSettings.externalSocksPort)
-            any = true
-        }
-        if (onionRelaysViaTor.value != torSettings.onionRelaysViaTor) {
-            onionRelaysViaTor.tryEmit(torSettings.onionRelaysViaTor)
-            any = true
-        }
-        if (dmRelaysViaTor.value != torSettings.dmRelaysViaTor) {
-            dmRelaysViaTor.tryEmit(torSettings.dmRelaysViaTor)
-            any = true
-        }
-        if (newRelaysViaTor.value != torSettings.newRelaysViaTor) {
-            newRelaysViaTor.tryEmit(torSettings.newRelaysViaTor)
-            any = true
-        }
-        if (trustedRelaysViaTor.value != torSettings.trustedRelaysViaTor) {
-            trustedRelaysViaTor.tryEmit(torSettings.trustedRelaysViaTor)
-            any = true
-        }
-        if (urlPreviewsViaTor.value != torSettings.urlPreviewsViaTor) {
-            urlPreviewsViaTor.tryEmit(torSettings.urlPreviewsViaTor)
-            any = true
-        }
-        if (profilePicsViaTor.value != torSettings.profilePicsViaTor) {
-            profilePicsViaTor.tryEmit(torSettings.profilePicsViaTor)
-            any = true
-        }
-        if (imagesViaTor.value != torSettings.imagesViaTor) {
-            imagesViaTor.tryEmit(torSettings.imagesViaTor)
-            any = true
-        }
-        if (videosViaTor.value != torSettings.videosViaTor) {
-            videosViaTor.tryEmit(torSettings.videosViaTor)
-            any = true
-        }
-        if (moneyOperationsViaTor.value != torSettings.moneyOperationsViaTor) {
-            moneyOperationsViaTor.tryEmit(torSettings.moneyOperationsViaTor)
-            any = true
-        }
-
-        if (nip05VerificationsViaTor.value != torSettings.nip05VerificationsViaTor) {
-            nip05VerificationsViaTor.tryEmit(torSettings.nip05VerificationsViaTor)
-            any = true
-        }
-        if (mediaUploadsViaTor.value != torSettings.mediaUploadsViaTor) {
-            mediaUploadsViaTor.tryEmit(torSettings.mediaUploadsViaTor)
-            any = true
-        }
-
-        return any
-    }
-
-    companion object {
-        fun build(torSettings: TorSettings): TorSettingsFlow =
-            TorSettingsFlow(
-                MutableStateFlow(torSettings.torType),
-                MutableStateFlow(torSettings.externalSocksPort),
-                MutableStateFlow(torSettings.onionRelaysViaTor),
-                MutableStateFlow(torSettings.dmRelaysViaTor),
-                MutableStateFlow(torSettings.newRelaysViaTor),
-                MutableStateFlow(torSettings.trustedRelaysViaTor),
-                MutableStateFlow(torSettings.urlPreviewsViaTor),
-                MutableStateFlow(torSettings.profilePicsViaTor),
-                MutableStateFlow(torSettings.imagesViaTor),
-                MutableStateFlow(torSettings.videosViaTor),
-                MutableStateFlow(torSettings.moneyOperationsViaTor),
-                MutableStateFlow(torSettings.nip05VerificationsViaTor),
-                MutableStateFlow(torSettings.mediaUploadsViaTor),
-            )
-    }
-}
+// Re-export from commons for backward compatibility
+typealias TorSettingsFlow = com.vitorpamplona.amethyst.commons.tor.TorSettingsFlow
